// src/types/colors.ts
export type ColorName =
  | 'green'
  | 'blue'
  | 'yellow'
  | 'red'
  | 'purple'
  | 'orange'
  | 'gray'
  | 'pink'
  | 'teal'
  | 'black'

export const COLOR_HEX: Record<ColorName, string> = {
  green: '#16a34a',
  blue: '#3b82f6',
  yellow: '#eab308',
  red: '#ef4444',
  purple: '#a855f7',
  orange: '#f97316',
  gray: '#6b7280',
  pink: '#ec4899',
  teal: '#06b6d4',
  black: '#000000',
}